ditch alpm, implement curl_multi for downloading
This is a fairly large change, moving repo code to a separate file,
removing alpm linkage, and (re)adding direct curl linkage. Rewrite the
downloader code using curl's multi interface, which results in far
better performance, being able to juggle multiple sockets at once.

Owning the download code is beneficial as it allows better control over
the downloader's behavior. This implements 304 response handling,
causing up to date repos to be skipped. We also hold the entirety of the
download in memory before committing to disk as an uncompressed CPIO
archive, avoiding the need for temporary files. However, similar to
pacman, the notion of continuing a repo download is ignored, since the
DB files are subject to a change in content without a change in name.