align output when writing to a terminal
This is a somewhat invasive change, breaking each line of a result into
a new struct, deemed 'struct line_t'. As items are added to the result
set, the length of the longest prefix is recorded for later output.

A new flag, -w/--raw is added to turn off this behavior in case there's
a need to parse the output.