Rename 'err' to 'dl_result', treat as an enum
04e42174dbad619237793d0ab174c1966dd04871 (years ago) sets repo->err=1
when the repo is up to date. I guess I did that so that 304s aren't
considered downloads. So, then, e1178fd508cb06a05c92383f47bb0652339d0d95
(new in v20) is the wrong fix because "up to date" is then considered an
error by the aggregation code (switching from case to an if/else). It
should have changed the real error cases to use repo->err=-1.

Fixes #50.