Replace fork() with threads for repacking repos
I'd like to think i've been lucky to use fork() without exec here, but
that might be too generous. A combination of std::async and std::future
makes this much nicer, and actually follows some semblance of rules now.

An interesting side effect of this change is that inode numbering
changes in the archives. libarchive seems to have some sort of global
inode counter that gets reset on a brand new process, but not on new pid
from a thread.