Overhaul filtering into its own mini-lib
Use some composition of classes with a tiny filter "language" to build
out the right combo of predicates to be used by the list/search
functions. This entirely supersedes the logic in match.hh and avoids
passing around state (largely, flags) in favor of better encapsulation.

This commit also introduces optional dependencies on gmock/gtest for
running unit tests (and ideally, soon to be more).