pkgfiled: an inotify-based daemon to sync files DBs
Since pacman 5, there's better support for .files databases, and pacman
has similar functionality to pkgfile. Instead of downloading the DBs
ourselves, and for the sake of consistency with pacman, transcode the
files DBs that pacman downloads via 'pacman -Fy' into the pkgfile cache
dir.

This may or may not resemble a daemon I've been using for 3+ years
written in Go.