share archive reader logic with main pkgfile code
A non-functional but possibly performance-related change here is that we
no longer mmap our repo file in favor of a bit of code simplicity. Share
the ArchiveReader and ReadOnlyFile we recently wrote with the
repo-loading logic for simplicity and to fix handling of the error path
when opening the repo+archive. No more evidence of goto!

This change also makes our path through repo-loading totally const, so
pass around const& repo_t, rather than mutable pointers.