Discover repos through directory traversal
This leaves ALPM config parsing (pacman.conf) solely to the update
logic, which I'd like to get rid of eventually. Doing so gives us a few
advantages:

1) No more need to build paths in ProcessRepo -- we already have the
path.
2) No more weird heuristics for determining if we found repos or not. We
know exactly what's in the cache directory.

Bonus: I no longer have to feel bad about never splitting out the warts
that grew on the repo class for the purposes of downloads because now
only the update code uses the Repo class.

main() gets a lot simpler as a result of this change and we capitalize
on further knock-on simplifications as a result of this.

NB: This *does* mean that results will no longer be in "repo order", but
we never actually advertised this as being a feature and it's hard to
believe that someone would depend on this (https://xkcd.com/1172/).