archive_reader: dramatically reduce number of string copies
This reduces the amount of data we copy out of the zero-copy buffer
provided to us by libarchive. The observation here is that search
operations (the more common operation) probably use less than 1% of the
overall data because they're returning a couple of lines out of maybe a
few archive entries.

Based on a query of 'pkgfile uv', valgrind shows:

Before:

  ==3794== HEAP SUMMARY:
  ==3794==     in use at exit: 0 bytes in 0 blocks
  ==3794==   total heap usage: 53,068 allocs, 53,068 frees, 3,253,624

After:

  ==3776== HEAP SUMMARY:
  ==3776==     in use at exit: 0 bytes in 0 blocks
  ==3776==   total heap usage: 14,617 allocs, 14,617 frees, 603,110 bytes

Implements a suggestion from #71.